common.utils: automatic refactor: bot decorators
Move `mod_only`, `sub_only`, `public_only`, `twitch_throttle` and
`throttle` to `lrrbot.decorators`.
