+++ lrrbot/commands/card.py
@@ -1,6 +1,7 @@
 import json
 import re
 
+import lrrbot.decorators
 from common import utils
 from lrrbot.main import bot
 
@@ -13,7 +14,7 @@ with open("mtgcards.json") as fp:
 	CARD_DATA = json.load(fp)
 
 @bot.command("card (.+)")
-@utils.throttle(60, count=3)
+@lrrbot.decorators.throttle(60, count=3)
 def card_lookup(lrrbot, conn, event, respond_to, search):
 	"""
 	Command: !card card-name
