+++ lrrbot/commands/explain.py
@@ -1,12 +1,13 @@
 import random
 
+import lrrbot.decorators
 from common import utils
 from lrrbot import storage
 from lrrbot.main import bot, log
 
 
 @bot.command("explain (.*?)")
-@utils.throttle(30, params=[4], count=2)
+@lrrbot.decorators.throttle(30, params=[4], count=2)
 def explain_response(lrrbot, conn, event, respond_to, command):
 	"""
 	Command: !explain TOPIC
